Modified order to E2fsprogs.xs file.
Libblkid API functions now appear in order in which they were added to the
libblkid library by version (oldest to most recent).

C preprocessor directives added to provide for conditional compilation so that
targeted builds of this package can be built to match the API of the installed
version of libblkid.