Updates to Makefile.PL, E2fsprogs.xs and inc/Devel/CheckLib.pm to support new interactive library detection configuration and installation process.
The package now uses an an interactive session with the user which
provides them with the choice to select dynamic library version detection
or to manually select the version of libblkid which they wish to
install on their systems. This is primarily driven via additions
to Makefile.PL, but is supported with new C preprocessor defines in the
xs file as well as supporting code in the CheckLib.pm module.